add vscode rest client http files
These files simplifiy the process of making a network request to the server in order to start scraping numbers and one to save these numbers.

It would be nice if the scraper fetched number periodically once it has all of them. Maybe look at the last line of the csv or just fetch whatever results are available every 5 minutes.